fix: Reimplement browser group key persistence on top of idb.
